Disable raw mode at exit and updates Makefile
Restore the terminal's original attributes when the program exits.
We'll save a copy of the termios struct in its original state,
and use tcsetattr() to apply it to the terminal when the program
exits.

atexit() comes from <stdlib.h> and is used to register disablingRawMode()
to be called automatically when the program exits. The original terminal is
stored in a global variable "orig_termios" and then is assigned in to the
raw struct to make a copy before we start making any changes.

Canonical mode is turned off, "ICANON" flag that allows us to turn off cano-
nical mode. Meaning that we will finally be reading byte-by-byte, instead
of line-by-line.

Signed-off-by: Francisco Pedraza <pacopedraza87@gmail.com>