Refactors keyboard input
Removed code from main function and call a function in main
instead.
editorReadKey() usage is to wait for one keypress, and return it.
We could expand this to handle escape sequences.
editorProcessKeyPress() waits for a keypress, and then handles it.
It can map various Ctrl key combinations and other special keys
to different editor functions.

Signed-off-by: Francisco Pedraza <pacopedraza87@gmail.com>