Adds editor function
write() and STDOUT_FILENO come from <unistd.h>
The 4 in our write() call means we are writing 4 bytes out to
the terminal. The first byte is "\x1b", which is the escape char-
acter, or 27 in decimal. The other three bytes are "[2J".

Escape Sequence always start with an escape character (27) followed
by a "[". those instruct the terminal to do various text formating
tasks, such as coloring text, moving the cursor around, and clearing
parts of the screen.

"J" command (Erase in Display) to clear the screen.

Signed-off-by: Francisco Pedraza <pacopedraza87@gmail.com>